# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Campus Platform API
  version: 1.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/activities:
    get:
      operationId: ListActivities
      tags: [activity]
      summary: "用户侧浏览已发布活动"
      x-generated-module: activity
      x-permission: activity:list
      security: [{bearerAuth: []}]
      parameters:
        - name: keyword
          in: query
          required: false
          schema: {type: string}
        - name: start_date
          in: query
          required: false
          schema: {type: string, format: date}
        - name: page
          in: query
          required: false
          schema: {type: integer, format: int32}
        - name: page_size
          in: query
          required: false
          schema: {type: integer, format: int32}
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /api/v1/activities/registrations/mine:
    get:
      operationId: ListMyActivityRegistrations
      tags: [activity]
      summary: "查看我的活动报名"
      x-generated-module: activity
      x-permission: activity:registration_mine
      security: [{bearerAuth: []}]
      parameters:
        - name: page
          in: query
          required: false
          schema: {type: integer, format: int32}
        - name: page_size
          in: query
          required: false
          schema: {type: integer, format: int32}
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /api/v1/activities/{id}:
    get:
      operationId: GetActivity
      tags: [activity]
      summary: "用户侧查看活动详情"
      x-generated-module: activity
      x-permission: activity:read
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/activities/{id}/registrations:
    post:
      operationId: CreateActivityRegistration
      tags: [activity]
      summary: "报名活动"
      x-generated-module: activity
      x-permission: activity:register
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/activities/{id}/registrations/me:
    delete:
      operationId: CancelMyActivityRegistration
      tags: [activity]
      summary: "取消我的活动报名"
      x-generated-module: activity
      x-permission: activity:registration_cancel
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expected_version
              properties:
                expected_version:
                  type: integer
                  format: uint64
                  minimum: 1
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/admin/activities:
    get:
      operationId: ListAdminActivities
      tags: [activity]
      summary: "管理端查询活动列表"
      x-generated-module: activity
      x-permission: activity:admin_list
      security: [{bearerAuth: []}]
      parameters:
        - name: keyword
          in: query
          required: false
          schema: {type: string}
        - name: status
          in: query
          required: false
          schema: {type: string}
        - name: review_status
          in: query
          required: false
          schema: {type: string}
        - name: start_date
          in: query
          required: false
          schema: {type: string, format: date}
        - name: page
          in: query
          required: false
          schema: {type: integer, format: int32}
        - name: page_size
          in: query
          required: false
          schema: {type: integer, format: int32}
      responses:
        "200":
          $ref: "#/components/responses/Success"
    post:
      operationId: CreateAdminActivity
      tags: [activity]
      summary: "管理端创建活动草稿"
      x-generated-module: activity
      x-permission: activity:admin_create
      security: [{bearerAuth: []}]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - summary
                - body
                - location
                - signup_start_at
                - signup_end_at
                - start_at
                - end_at
                - capacity
                - contact_type
                - contact
              properties:
                title:
                  type: string
                summary:
                  type: string
                body:
                  type: string
                location:
                  type: string
                signup_start_at:
                  type: string
                  format: date-time
                signup_end_at:
                  type: string
                  format: date-time
                start_at:
                  type: string
                  format: date-time
                end_at:
                  type: string
                  format: date-time
                capacity:
                  type: integer
                  format: int64
                  minimum: 1
                contact_type:
                  type: string
                contact:
                  type: string
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/admin/activities/{id}:
    get:
      operationId: GetAdminActivity
      tags: [activity]
      summary: "管理端查看活动详情"
      x-generated-module: activity
      x-permission: activity:admin_read
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      operationId: UpdateAdminActivity
      tags: [activity]
      summary: "管理端修改活动草稿"
      x-generated-module: activity
      x-permission: activity:admin_update
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - summary
                - body
                - location
                - signup_start_at
                - signup_end_at
                - start_at
                - end_at
                - capacity
                - expected_version
              properties:
                title:
                  type: string
                summary:
                  type: string
                body:
                  type: string
                location:
                  type: string
                signup_start_at:
                  type: string
                  format: date-time
                signup_end_at:
                  type: string
                  format: date-time
                start_at:
                  type: string
                  format: date-time
                end_at:
                  type: string
                  format: date-time
                capacity:
                  type: integer
                  format: int64
                  minimum: 1
                contact_type:
                  type: string
                contact:
                  type: string
                expected_version:
                  type: integer
                  format: uint64
                  minimum: 1
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/admin/activities/{id}/approve:
    post:
      operationId: ApproveAdminActivity
      tags: [activity]
      summary: "审核通过活动"
      x-generated-module: activity
      x-permission: activity:admin_approve
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expected_version
              properties:
                expected_version:
                  type: integer
                  format: uint64
                  minimum: 1
                review_comment:
                  type: string
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/admin/activities/{id}/cancel:
    post:
      operationId: CancelAdminActivity
      tags: [activity]
      summary: "取消活动"
      x-generated-module: activity
      x-permission: activity:admin_cancel
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expected_version
              properties:
                expected_version:
                  type: integer
                  format: uint64
                  minimum: 1
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/admin/activities/{id}/finish:
    post:
      operationId: FinishAdminActivity
      tags: [activity]
      summary: "结束活动"
      x-generated-module: activity
      x-permission: activity:admin_finish
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expected_version
              properties:
                expected_version:
                  type: integer
                  format: uint64
                  minimum: 1
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/admin/activities/{id}/publish:
    post:
      operationId: PublishAdminActivity
      tags: [activity]
      summary: "发布活动"
      x-generated-module: activity
      x-permission: activity:admin_publish
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expected_version
              properties:
                expected_version:
                  type: integer
                  format: uint64
                  minimum: 1
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/admin/activities/{id}/reject:
    post:
      operationId: RejectAdminActivity
      tags: [activity]
      summary: "驳回活动审核"
      x-generated-module: activity
      x-permission: activity:admin_reject
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expected_version
                - review_comment
              properties:
                expected_version:
                  type: integer
                  format: uint64
                  minimum: 1
                review_comment:
                  type: string
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/admin/activities/{id}/submit-review:
    post:
      operationId: SubmitAdminActivityReview
      tags: [activity]
      summary: "提交活动审核"
      x-generated-module: activity
      x-permission: activity:admin_submit_review
      security: [{bearerAuth: []}]
      parameters:
        - {name: id, in: path, required: true, schema: {type: integer, format: uint64, minimum: 1}}
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - expected_version
              properties:
                expected_version:
                  type: integer
                  format: uint64
                  minimum: 1
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
...
  responses:
    Success:
      description: Successful response
      content: {application/json: {schema: {$ref: "#/components/schemas/SuccessEnvelope"}}}
...
    Error:
      description: Error response
      content: {application/json: {schema: {$ref: "#/components/schemas/ErrorEnvelope"}}}
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [data, request_id]
      properties: {data: {}, request_id: {type: string}}
...
    ErrorEnvelope:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties: {code: {type: string}, message: {type: string}}
        request_id: {type: string}